# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Coffee Distribution CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: active
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: customerId
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Not found
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update an order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        customerChannel:
          type: string
        segment:
          type: string
        priorityClass:
          type: string
          enum: [A, B, C]
        paymentTerms:
          type: string
        discountLevel:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        active:
          type: boolean
        businessChannel:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - companyName
        - contactPerson
        - phone
        - email
        - customerChannel
        - segment
        - priorityClass
        - paymentTerms
        - active
        - businessChannel
        - createdAt
        - updatedAt
...
    Product:
      type: object
      properties:
        id:
          type: integer
        productName:
          type: string
        sku:
          type: string
        category:
          type: string
        unitPrice:
          type: number
        stockStatus:
          type: string
          enum: [in_stock, low_stock, out_of_stock]
        active:
          type: boolean
        businessChannel:
          type: string
        createdAt:
          type: string
      required:
        - id
        - productName
        - sku
        - category
        - unitPrice
        - stockStatus
        - active
        - businessChannel
        - createdAt
    CreateProductBody:
      type: object
      properties:
        productName:
          type: string
        sku:
          type: string
        category:
          type: string
        unitPrice:
          type: number
        stockStatus:
          type: string
          enum: [in_stock, low_stock, out_of_stock]
        active:
          type: boolean
        businessChannel:
          type: string
      required:
        - productName
        - sku
        - category
        - unitPrice
        - stockStatus
        - businessChannel
    UpdateProductBody:
      type: object
      properties:
        productName:
          type: string
        sku:
          type: string
        category:
          type: string
        unitPrice:
          type: number
        stockStatus:
          type: string
          enum: [in_stock, low_stock, out_of_stock]
        active:
          type: boolean
        businessChannel:
          type: string
    Order:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        businessChannel:
          type: string
        orderSource:
          type: string
          enum: [phone, whatsapp, web, sales_rep]
        requestedDeliveryDate:
          type: ["string", "null"]
        urgency:
          type: string
          enum: [low, normal, high, critical]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, confirmed, in_progress, delivered, cancelled]
        totalAmount:
          type: number
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - customerId
        - customerName
        - businessChannel
        - orderSource
        - urgency
        - status
        - totalAmount
        - createdAt
        - updatedAt
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
            customer:
              $ref: "#/components/schemas/Customer"
          required:
            - items
            - customer
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPriceSnapshot:
          type: number
        lineTotal:
          type: number
      required:
        - id
        - orderId
        - productId
        - productName
        - quantity
        - unitPriceSnapshot
        - lineTotal
    CreateOrderBody:
      type: object
      properties:
        customerId:
          type: integer
        businessChannel:
          type: string
        orderSource:
          type: string
          enum: [phone, whatsapp, web, sales_rep]
        requestedDeliveryDate:
          type: ["string", "null"]
        urgency:
          type: string
          enum: [low, normal, high, critical]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemBody"
      required:
        - customerId
        - businessChannel
        - orderSource
        - urgency
        - items
    CreateOrderItemBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        unitPriceSnapshot:
          type: number
      required:
        - productId
        - quantity
        - unitPriceSnapshot
    UpdateOrderBody:
      type: object
      properties:
        status:
          type: string
          enum: [draft, confirmed, in_progress, delivered, cancelled]
        requestedDeliveryDate:
          type: ["string", "null"]
        urgency:
          type: string
          enum: [low, normal, high, critical]
        notes:
          type: ["string", "null"]
...